requirements.txt
# GENERATED FROM requirements.in.  DO NOT EDIT DIRECTLY.
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.1
celery-redbeat==2.3.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Flask==3.1.0
Flask-Dance==7.1.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.23.3
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
MarkupSafe==3.0.2
oauthlib==3.2.2
openai==1.75.0
packaging==24.2
pdf2image==1.17.0
pillow==11.2.1
pipdeptree==2.26.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
URLObject==2.4.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yapf==0.43.0
